OPT-988 move map projection contracts
