+++ src/app/controller/ui/starmap_view/mod.rs
@@ -7,4 +7,4 @@ mod models;
 mod repository;
 
 pub(crate) use jobs::{run_umap_build, run_umap_cluster_build};
-pub(crate) use models::{UmapBounds, UmapPoint, UmapPointQuery};
+pub(crate) use models::{UmapBounds, UmapPoint};
