+++ src/app/controller/ui/starmap_view/models.rs
@@ -1,7 +1,5 @@
 //! Shared starmap data models used by the UI controller and repository loaders.
 
-use crate::sample_sources::SourceId;
-
 /// Aggregate starmap bounds for the current layout.
 pub(crate) struct UmapBounds {
     pub min_x: f32,
@@ -16,15 +14,4 @@ pub(crate) struct UmapPoint {
     pub x: f32,
     pub y: f32,
     pub cluster_id: Option<i32>,
-}
-
-/// Query payload for loading visible starmap points and optional cluster metadata.
-pub(crate) struct UmapPointQuery<'a> {
-    pub model_id: &'a str,
-    pub umap_version: &'a str,
-    pub cluster_method: &'a str,
-    pub cluster_umap_version: &'a str,
-    pub source_id: Option<&'a SourceId>,
-    pub bounds: crate::app::state::MapQueryBounds,
-    pub limit: usize,
 }
