+++ src/app_core/app_api.rs
@@ -10,7 +10,7 @@
 //! | --- | --- | --- | --- | --- |
 //! | `AppController`, startup/test fixture helpers, and WAV edit-support predicate | `src/app::controller` | Intentionally still owned by `src/app` while mature runtime behavior remains there | `OPT-992` | App-core exposes a narrow runtime facade; fixture construction is test-owned; domain helpers no longer require broad controller aliases |
 //! | Browser projection cache, selected-path lookup, preload-window, and auto-rename row state | `src/app_core::browser_projection_cache` | App-core owned as of `OPT-987` | Done | Remaining work should keep new browser projection contracts in app-core and avoid reintroducing bridge aliases |
-//! | Map projection cache key, projected map-point entry, and UMAP point query payload | `src/app::controller` retained projection/query state | Ready for app-core ownership | `OPT-988` | Map projection consumes app-core-owned query/cache contracts or a narrow map runtime adapter |
+//! | Map projection cache key, projected map-point entry, and UMAP point query payload | `src/app_core::map_projection_contracts` | App-core owned as of `OPT-988` | Done | Remaining work should keep new map projection contracts in app-core and avoid reintroducing bridge aliases |
 //! | Dirty graph node/reason aliases used by frame preparation and invalidation adapters | `src/app::controller::state::runtime` | Ready for app-core invalidation ownership | `OPT-989` | App-core frame preparation and bridge invalidation use app-core invalidation names; legacy conversion is isolated or removed |
 //! | Browser, source, folder, and library-hygiene state DTOs exposed through the wildcard state bridge | `src/app::state` | Ready for app-core state ownership | `OPT-990` | Browser/source/folder projections and tests use app-core DTOs or focused builders |
 //! | Waveform, prompt, drag/drop, map, audio/options, progress, update, and status state DTOs exposed through the wildcard state bridge | `src/app::state` | Ready for app-core state ownership | `OPT-991` | Covered projection/action consumers use app-core DTOs or focused builders |
@@ -23,8 +23,7 @@
 
 pub(crate) mod controller {
     pub(crate) use crate::app::controller::{
-        AppController, ProjectedMapPointCacheEntry, ProjectedMapPointsCacheKey, UmapPointQuery,
-        build_named_gui_fixture_controller, supports_wav_destructive_edits,
+        AppController, build_named_gui_fixture_controller, supports_wav_destructive_edits,
     };
 
     pub(crate) type DerivedNodeId = crate::app::controller::state::runtime::DerivedNodeId;
