+++ src/app_core/mod.rs
@@ -7,6 +7,9 @@ mod app_api;
 /// Retained browser projection cache contracts owned by app-core.
 pub(crate) mod browser_projection_cache;
 
+/// Retained map projection query/cache contracts owned by app-core.
+pub(crate) mod map_projection_contracts;
+
 /// Controller aliases and helpers used by migration-facing runtimes.
 pub mod controller;
